[경로 이동][BOJ] 수 정렬하기 1 / 2750 / 브론즈1 / crtEvent
https://www.acmicpc.net/problem/2750

related to #21
