+++ crtEvent/doit/ch04_정렬/sec01_버블정렬/p2750_수정렬하기_1/Main02_BubbleSort_2.java
@@ -1,15 +1,12 @@
-package doit.ch04_정렬.p2750_수정렬하기_1;
+package doit.ch04_정렬.sec01_버블정렬.p2750_수정렬하기_1;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.util.Arrays;
 
-public class Main04_ArraysSort {
+public class Main02_BubbleSort_2 {
 
-    // Arrays.sort() 로 정렬
-    // 176ms
-
+    // 약간의 리팩토링
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
@@ -21,7 +18,15 @@ public class Main04_ArraysSort {
         }
 
         // 정렬
-        Arrays.sort(arr);
+        for(int i = 0; i < N-1; i++) {
+            for(int j = 0; j < N-1-i; j++) {
+                if(arr[j] > arr[j+1]) {
+                    int temp = arr[j];
+                    arr[j] = arr[j+1];
+                    arr[j+1] = temp;
+                }
+            }
+        }
 
         // 출력
         for(int i = 0; i < N; i++) {
@@ -29,5 +34,4 @@ public class Main04_ArraysSort {
         }
 
     }
-
 }
