+++ crtEvent/doit/ch04_정렬/sec01_버블정렬/p2750_수정렬하기_1/Main03_StreamSorted.java
@@ -1,0 +1,26 @@
+package doit.ch04_정렬.sec01_버블정렬.p2750_수정렬하기_1;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.util.Arrays;
+
+public class Main03_StreamSorted {
+
+    // Array.stream의 sorted() 메소드로 정렬
+    // 192ms
+
+    public static void main(String[] args) throws IOException {
+        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+
+        int  N = Integer.parseInt(br.readLine());
+
+        int[] arr = new int[N];
+        for(int i = 0; i < N; i++) {
+            arr[i] = Integer.parseInt(br.readLine());
+        }
+
+        Arrays.stream(arr).sorted().forEach(System.out::println);
+
+    }
+}
